(GH-459) Consolidate debug features into a single feature
Previously there were two debug features, one for the factory and one
for the configuration.  This commit moves the functionality of the
DebugConfiguration into the DebuggingFeature.
